Split user logic between MVC and a new middleware module
