+++ src/controllers/user/mod.rs
@@ -1,0 +1,3 @@
+pub mod me;
+pub mod session;
+pub mod other;
