+++ src/controllers/user/session.rs
@@ -1,0 +1,124 @@
+use controllers::prelude::*;
+
+use app::RequestApp;
+use conduit_cookie::RequestSession;
+use rand::{thread_rng, Rng};
+use github;
+
+use models::NewUser;
+
+/// Handles the `GET /authorize_url` route.
+///
+/// This route will return an authorization URL for the GitHub OAuth flow including the crates.io
+/// `client_id` and a randomly generated `state` secret.
+///
+/// see <https://developer.github.com/v3/oauth/#redirect-users-to-request-github-access>
+///
+/// ## Response Body Example
+///
+/// ```json
+/// {
+///     "state": "b84a63c4ea3fcb4ac84",
+///     "url": "https://github.com/login/oauth/authorize?client_id=...&state=...&scope=read%3Aorg"
+/// }
+/// ```
+pub fn github_authorize(req: &mut Request) -> CargoResult<Response> {
+    // Generate a random 16 char ASCII string
+    let state: String = thread_rng().gen_ascii_chars().take(16).collect();
+    req.session()
+        .insert("github_oauth_state".to_string(), state.clone());
+
+    let url = req.app().github.authorize_url(state.clone());
+
+    #[derive(Serialize)]
+    struct R {
+        url: String,
+        state: String,
+    }
+    Ok(req.json(&R {
+        url: url.to_string(),
+        state: state,
+    }))
+}
+
+/// Handles the `GET /authorize` route.
+///
+/// This route is called from the GitHub API OAuth flow after the user accepted or rejected
+/// the data access permissions. It will check the `state` parameter and then call the GitHub API
+/// to exchange the temporary `code` for an API token. The API token is returned together with
+/// the corresponding user information.
+///
+/// see <https://developer.github.com/v3/oauth/#github-redirects-back-to-your-site>
+///
+/// ## Query Parameters
+///
+/// - `code` – temporary code received from the GitHub API  **(Required)**
+/// - `state` – state parameter received from the GitHub API  **(Required)**
+///
+/// ## Response Body Example
+///
+/// ```json
+/// {
+///     "api_token": "b84a63c4ea3fcb4ac84",
+///     "user": {
+///         "email": "foo@bar.org",
+///         "name": "Foo Bar",
+///         "login": "foobar",
+///         "avatar": "https://avatars.githubusercontent.com/u/1234",
+///         "url": null
+///     }
+/// }
+/// ```
+pub fn github_access_token(req: &mut Request) -> CargoResult<Response> {
+    // Parse the url query
+    let mut query = req.query();
+    let code = query.remove("code").unwrap_or_default();
+    let state = query.remove("state").unwrap_or_default();
+
+    // Make sure that the state we just got matches the session state that we
+    // should have issued earlier.
+    {
+        let session_state = req.session().remove(&"github_oauth_state".to_string());
+        let session_state = session_state.as_ref().map(|a| &a[..]);
+        if Some(&state[..]) != session_state {
+            return Err(human("invalid state parameter"));
+        }
+    }
+
+    #[derive(Deserialize)]
+    struct GithubUser {
+        email: Option<String>,
+        name: Option<String>,
+        login: String,
+        id: i32,
+        avatar_url: Option<String>,
+    }
+
+    // Fetch the access token from github using the code we just got
+    let token = req.app()
+        .github
+        .exchange(code.clone())
+        .map_err(|s| human(&s))?;
+
+    let (handle, resp) = github::github(req.app(), "/user", &token)?;
+    let ghuser: GithubUser = github::parse_github_response(handle, &resp)?;
+
+    let user = NewUser::new(
+        ghuser.id,
+        &ghuser.login,
+        ghuser.email.as_ref().map(|s| &s[..]),
+        ghuser.name.as_ref().map(|s| &s[..]),
+        ghuser.avatar_url.as_ref().map(|s| &s[..]),
+        &token.access_token,
+    ).create_or_update(&*req.db_conn()?)?;
+    req.session()
+        .insert("user_id".to_string(), user.id.to_string());
+    req.mut_extensions().insert(user);
+    super::me::me(req)
+}
+
+/// Handles the `GET /logout` route.
+pub fn logout(req: &mut Request) -> CargoResult<Response> {
+    req.session().remove(&"user_id".to_string());
+    Ok(req.json(&true))
+}
