+++ src/middleware/mod.rs
@@ -1,0 +1,3 @@
+pub use self::current_user::CurrentUser;
+
+pub mod current_user;
