+++ src/models/email.rs
@@ -1,0 +1,22 @@
+use chrono::NaiveDateTime;
+
+use models::User;
+use schema::emails;
+
+#[derive(Debug, Queryable, AsChangeset, Identifiable, Associations)]
+#[belongs_to(User)]
+pub struct Email {
+    pub id: i32,
+    pub user_id: i32,
+    pub email: String,
+    pub verified: bool,
+    pub token: String,
+    pub token_generated_at: Option<NaiveDateTime>,
+}
+
+#[derive(Debug, Insertable, AsChangeset)]
+#[table_name = "emails"]
+pub struct NewEmail<'a> {
+    pub user_id: i32,
+    pub email: &'a str,
+}
