+++ src/models/mod.rs
@@ -3,13 +3,14 @@ pub use self::category::{Category, CrateCategory, NewCategory};
 pub use self::crate_owner_invitation::{CrateOwnerInvitation, NewCrateOwnerInvitation};
 pub use self::dependency::{Dependency, DependencyKind, ReverseDependency};
 pub use self::download::VersionDownload;
+pub use self::email::{Email, NewEmail};
 pub use self::follow::Follow;
 pub use self::keyword::{CrateKeyword, Keyword};
 pub use self::krate::{Crate, CrateDownload, NewCrate};
 pub use self::owner::{CrateOwner, Owner, OwnerKind};
 pub use self::rights::Rights;
 pub use self::team::{NewTeam, Team};
-pub use user::{Email, NewUser, User};
+pub use self::user::{NewUser, User};
 pub use self::token::ApiToken;
 pub use self::version::{NewVersion, Version};
 
@@ -20,6 +21,7 @@ mod category;
 mod crate_owner_invitation;
 pub mod dependency;
 mod download;
+mod email;
 mod follow;
 mod keyword;
 pub mod krate;
@@ -27,4 +29,5 @@ mod owner;
 mod rights;
 mod team;
 mod token;
+mod user;
 mod version;
